Automated g4 rollback of changelist 194599395.
*** Reason for rollback ***

Rollforward now that cr/195108085 is in, which disables the test that was broken. Confirmed that the functionality isn't actually broken, and we suspect there are issues with the underlying test runner. See b/79150203 for more details.

*** Original change description ***

Automated g4 rollback of changelist 194576732.

*** Reason for rollback ***

*** Original change description ***

Update the KeyHandler and ActionHandler to use the keydown event data rather than the keypress event data to determine whether a key should be handled for Firefox.

Originally we used keypress for Gecko a...

***

RELNOTES: Update KeyHandler and ActionHandler to use keydown events as the source for KeyEvents when using Gecko, rather than keypress events. This was only necessary pre-Gecko 4. This brings the handling in line with other browsers and fixes https://github.com/google/closure-library/issues/883.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=196114041